requirements.txt
wheel==0.38.4
roslibpy==1.3.0
rich==12.6.0
requests==2.28.1
paramiko~=2.11.0
inputs~=0.5.0
PyQt5~=5.15.4
numpy~=1.23.4
Pillow~=9.2.0
opencv-python~=4.6.0.66
psutil~=5.9.3
humanize~=4.4.0